pyproject.toml
[project]
name = "grgriepreis"
version = "0.1.0"
description = "Lidl website scraper and product detector"
authors = [
    {name = "GrGrie",email = "gsgrechkin@gmail.com"}
]
license = {text = "GrGrie"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "numpy (>=2,<2.3.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "boto3 (>=1.40.26,<2.0.0)",
    "selenium (>=4.35.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "ultralytics (>=8.3.199,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]

[tool.poetry.group.dev.dependencies]
ultralytics = "^8.3.199"

